Converted EuiStep to TS
